changed name to scpi_psu and moved dependencies to /socs/agent
